# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ClippicErrorInterface": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32"
					},
					"trace": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"code",
					"trace"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetSurnameResponseData": {
				"properties": {
					"surname": {
						"type": "string",
						"description": "The user's surname",
						"example": "tester",
						"maxLength": 100
					}
				},
				"required": [
					"surname"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetSurnameResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetSurnameResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutSurnameResponseData": {
				"properties": {
					"surname": {
						"type": "string",
						"description": "The user's surname",
						"example": "tester",
						"maxLength": 100
					}
				},
				"required": [
					"surname"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutSurnameResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PutSurnameResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutSurnameRequest": {
				"description": "Create or update the users surname",
				"properties": {
					"surname": {
						"type": "string",
						"description": "The user's surname",
						"example": "tester",
						"minLength": 1,
						"maxLength": 100
					}
				},
				"required": [
					"surname"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"surname": "tester"
				}
			},
			"GetUsernameResponseData": {
				"properties": {
					"username": {
						"type": "string",
						"description": "The user's username",
						"example": "tester",
						"maxLength": 40
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetUsernameResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetUsernameResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutUsernameResponseData": {
				"properties": {
					"username": {
						"type": "string",
						"description": "The user's username",
						"example": "tester",
						"maxLength": 40
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutUsernameResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PutUsernameResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UsernameAlreadyExistsError": {
				"$ref": "#/components/schemas/ClippicErrorInterface",
				"description": "1048 | Error when username already exists."
			},
			"PutUsernameRequest": {
				"description": "Create or update the users username",
				"properties": {
					"username": {
						"type": "string",
						"description": "The user's username",
						"example": "tester",
						"minLength": 6,
						"maxLength": 40
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"username": "tester"
				}
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@clippic/user-v2",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users/v2/surname": {
			"get": {
				"operationId": "GetSurnameRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSurnameResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"surname": "tester"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will return the user's surname.",
				"tags": [
					"Surname"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutSurnameRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PutSurnameResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"surname": "tester"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will update or create if not exists the user's surname.",
				"tags": [
					"Surname"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutSurnameRequest"
							}
						}
					}
				}
			}
		},
		"/users/v2/username": {
			"get": {
				"operationId": "GetUsernameRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsernameResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"username": "tester"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will return the user's username.",
				"tags": [
					"Username"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutUsernameRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PutUsernameResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"username": "tester"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsernameAlreadyExistsError"
								}
							}
						}
					}
				},
				"description": "This request will update or create if not exists the user's username.\n\n**Errors:**\n\n| Code | Description                 |\n|------|-----------------------------|\n| 400  | UsernameAlreadyExistsError  |",
				"tags": [
					"Username"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutUsernameRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}